requirements.txt
asgiref==3.8.1
Django==5.0.6
pillow==10.3.0
#psycopg2==2.9.9
sqlparse==0.5.0
whitenoise==6.6.0
